+++ common/src/main/resources/architectury.accessWidener
@@ -97,6 +97,7 @@ transitive-accessible field net/minecraft/client/renderer/rendertype/RenderType 
 transitive-accessible field net/minecraft/client/renderer/rendertype/RenderType name Ljava/lang/String;
 accessible class net/minecraft/client/particle/ParticleResources$MutableSpriteSet
 accessible field net/minecraft/client/particle/ParticleResources$MutableSpriteSet sprites Ljava/util/List;
+accessible field net/minecraft/client/particle/ParticleEngine resourceManager Lnet/minecraft/client/particle/ParticleResources;
 transitive-accessible class net/minecraft/world/item/CreativeModeTab$Output
 transitive-accessible class net/minecraft/world/item/CreativeModeTab$TabVisibility
 accessible field net/minecraft/client/multiplayer/MultiPlayerGameMode connection Lnet/minecraft/client/multiplayer/ClientPacketListener;
