+++ fabric/build.gradle
@@ -25,7 +25,7 @@ repositories {
 dependencies {
     api "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
     api "net.fabricmc.fabric-api:fabric-api:${rootProject.fabric_api_version}"
-    //compileOnly("com.terraformersmc:modmenu:${rootProject.mod_menu_version}") { transitive false }
+    compileOnly("com.terraformersmc:modmenu:${rootProject.mod_menu_version}") { transitive false }
 
     common(project(path: ":common")) { transitive false }
     shadowCommon(project(path: ":common", configuration: "transformProductionFabric")) { transitive false }
