+++ fabric/src/main/resources/architectury.mixins.json
@@ -11,7 +11,7 @@
     "client.MixinClientLevel",
     "client.MixinClientPacketListener",
     "client.MixinFabricClientCommandSource",
-    "client.MixinGameRenderer",
+    "client.MixinGui",
     "client.MixinGuiGraphicsExtractor",
     "client.MixinIntegratedServer",
     "client.MixinKeyboardHandler",
