+++ gradle.properties
@@ -16,7 +16,7 @@ version_suffix=
 
 fabric_loader_version=0.19.3
 fabric_api_version=0.152.0+26.2
-#mod_menu_version=18.0.0-beta.1
+mod_menu_version=20.0.0-alpha.1
 
 neoforge_version=26.2.0-alpha.0+rc-2.20260612.124211
